requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
keyboard==0.13.5
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.137
marshmallow==3.23.0
MouseInfo==0.1.3
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.3.3
orjson==3.10.10
packaging==24.1
pillow==11.0.0
propcache==0.2.0
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
pytweening==1.2.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
